// modules/movie/interfaces/movie.interface.ts
export interface MovieInterface {
  id: number;
  character_id: number;
  character_level: number;
  guild_id: number;
  set: string;
  title: string;
  custom_title: boolean;
  cover_background_type: number;
  cover_border_type: number;
  cover_layout_type: number;
  cover_layout_new_type: number;
  cover_font_type: number;
  cover_actor_type: number;
  cover_filter_type: number;
  cover: string;
  status: number;
  rating: number;
  fans: number;
  energy: number;
  stat: number;
  needed_energy: number;
  duration: number;
  ts_end: number;
  movie_quest_id: number;
  movie_quest_pool: string;
  claimed_stars: number;
  rewards_star_1: string;
  rewards_star_2: string;
  rewards_star_3: string;
  votes: number;
  rank: number;
}